change grammar data structure into DSL data structure
As this library will be used downstream, the "weights" of
the grammar do not effect how a program is synthesized from
a particular selection of primitives.  Thus the name "grammar"
seems highly misleading; the only usage of this data structure
is as a container for a particular set of primitives (its library)
and a designation of the state type if this DSL belongs to a domain
that uses continuuation passing style with some base type representing
domain state.  In contexts where a likelihood or probability is
still being assigned to a primitive, this calculation will always
be based on a uniform distribution over all library entries, and thus
nothing but the size of the DSL is needed for these calculations;
no individualized probabilities need to be stored on the entry.