change command translation strategy; remove references to "grammar"
These changes replace all usages of "grammar" or "gmr" with "dsl"
throughout all modules in line with preceding commits.  They also
update the translation/search strategy used by the `Commands` module.

The goal of this module is to implement a `commands_to_program`
function that is total and whose image is the entire set of valid
programs constructed of `n_commands` or less primitives.  The
implementation added in this commit fulfills these two suggestions.
Since the preimage of this function (all valid commands sequences)
is almost certainly larger than its image (valid programs of size
<= `n_commands`), it is almost certainly surjective rather than
bijective.

The `commands_to_programs` function takes commands as input where
each command represents a `search_point`.  A `search_point` is a
fraction between zero and `(dsl.size - 1) / dsl.size`, inclusive,
which can be scaled and rounded to index any primitive from a list
`dsl.size` or fewer primitives.  `commands_to_program` performs a
depth-first search by assembling programs according to the search
points, where a hole in a partial program is filled by selecting
from the list of available unifiable primitives. `commands_to_program`
begins its search with the program assembled from the initial selections
given by the search points; if this does not specify a complete program
of size <= `n_commands`, it continues by attempting to fill the most
deeply nested hole by selecting according to the search point from
among the remaining available unifying primitives.  So a sequence
of commands can be understood as providing an ordering on valid
programs.  This ordering is deterministic but arbitrary/depends on
the arbitrary ordering of primitives within a DSL.  Unless the
"intended" program is the very first one as dictated by this ordering,
it may be hard to understand the resulting output.  Nonetheless, this
inscrutable behavior only occurs for otherwise invalid inputs, and
preserves the largest prefix of the flat sequence of primitives used
to generate the program.