update program execution return JSON
"translated" is now a clearer name for the field indicating the success/failure
of `commands_to_program` than `typechecked`.  Additionally, the "domain" keyword
argument is unnecessary, as anyone calling this functionality in the style
demonstrated in antireduce-graphs will be specializing the other arguments appropriately
and thus the caller will know the domain.