revise translation strategy
The translation strategy implemented by `commands_to_program`
was not succeeding at finding complete programs because
when enumerating a program of type `state_type -> state_type` from
a DSL which does not contain any entry of type `state_type`, the
use of at least one variable (the variable bound by the outermost
lambda of the enumerated program) is required to satisfy the type.

To address this, I decided to limit all domains built atop antireduce
to being in continuation passing style.  Thus, `req` as passed to
`enumerate_terminal` in the initial call is always of type `state_type
-> state_type`.  This allows the addition of a single optional argument
to trigger the appropriate insertion of `$0` at the one and only location
where the use of a variable is required, and to utilize the typical
recursive procedure based on search points in all other circumstances.

The only use of variables in lambda calculus programs is to make a larger
program more concise and have fewer exposed parameters.  But for the
uses this library is intended for, machine-crafted programs will almost
certainly have a uniform signature across different calls to these methods.
If I expect these signatures to be fixed, they may as well be limited to
accepting a single state which may be backed by a data structure supporting
the specific base DSL that those building on this library want to support.