make the state_type of a DSL non-optional; correct translation strategy
In line with previous changes, the `commands_to_program` function expects
a request of the form `state_type -> state_type`, so `state_type`s are not
really an optional attribute of a DSL; this was an accidental omission.

The translation strategy was still not functioning properly because
the use of a variable to supply a value of type `state_type` is not
necessary only in the outermost abstraction generated in response to
the initial request `state_type -> state_type`, but any time that
this request is generated recursively (which it often is).  The previous
changes were allowing one outermost lambda to be generated but were
causing the same problem---no complete program could be created without
some way to supply a value of type `state_type`---when trying to generate
anything but `identity $0` as its body.