supply search points as integers rather than floats
Previously, it was expected that search points came as floats
which had some uniform, maximum precision.  I realized that this
makes it harder, for humans and likely for machines, to generate
specific programs, because in addition to knowing the length
of the list of unifying primitives that a certain search point
is choosing from (and its contents), one has to know which float,
when multiplied by this length, yields an index which will round
to the index of the desired primitive.  This is more challenging
than a semantics where search points are supplied as integers,
and those integers are taken as specifying such indices directly.

Integer search point semantics change the way it is handled
when an earlier search point is "misspecified:" if a later search
point specifies an out-of-bounds index, then the search point is
taken to refer to the last index in the list of unifying primitives,
whatever its length, rather than drawing from a point some specified
fraction of the way through this list.  I am not sure that either
of these behaviors is more interpretable than the other in the
case of misspecifications, but the integer search point semantics
are certainly easier in the absence of misspecifications, since
for `0 <= i < n_unifying_primitives <= j <= (dsl.size - 1)` each `i`
selects a distinct primitive and each `j` selects the same primitive.
With the floating point semantics, the search points that do not
specify different underlying primitives are spaced out at various
locations between `0` and `dsl.size - 1`, depending on
`n_unifying_primitives`.  So with the integer semantics, a human
or machine user picks the primitives it thinks it wants exactly,
based on its understanding of the unifying primitives at each stage
of program generation; if these are not quite what was anticipated,
then certain parts of the program may be filled in with very hard to
predict (but deterministic) ways.  With the floating point semantics,
a human has to pick the floating point that, when discretized,
selects the primitive it thinks it wants exactly; if these are
not what it thinks they are, the remainder of the program will
still be filled in in unpredictable (but deterministic) ways but
with some attempt to preserve the "intent" of the search points by
using them to select a proportion of the way through the list of
unifying primitives at which search begins rather than an exact
position.  I think that the complexity of the latter option is not
worth the cost.