correct overagressive optimization in `commands_to_program`
If there are not at least as many search points remaining as there
are parameters needed to fully apply the current primitive at a
given point in the translation process of `commands_to_program`, then
we know that the program cannot be completed using that primitive.
This optimization was kept the same even as changes were made to
accomodate the fact that final `state_type` parameter is always
and necessarily supplied as `$0` (the most tightly bound variable)
and therefore does not consume a search point.  The function used to
calculate the number of parameters needed to fully apply a primitive
was updated so as to incorporate this understanding.