remove check for domain of executed program
It is now assumed that the contents of the directory passed are all
executed program JSON (due to the tacit assumption that domain data
is kept in a domain-specific location).  As a result, the name of
the domain is not stored in executed program JSON and the corresponding
field no longer exists.