address underflow in handling of normalization factors
When calculating the likelihood from a factorization, the logsumexp operation
used to find the approximate maximum likelihood of the primitives in the denominator
was experiencing numerical instabilities.  A formula that wikipedia states is more
numerically stable has been substituted.  The LSE is now calculated as:

    x* + log(\sum_i exp(x_i - x*))

where x* = max_i x_i, in place of

    log(\sum_i log x_i).