correct incorrect input to rewriting procedure
Due to an inconsistency accidentally introduced when refurbishing the dreamcoder
codebase, the rewriting procedure was called such that it sought to replace instances
of `Invented refactoring` rather than `refactoring` in the inverse beta-reduced programs.
When this was not found (since the programs are not yet rewritten with `refactoring` as
an invented primitive), the inverse beta-reduced program resembled an ordinary program which
was out of beta-normal form.  Imitating the original dreamcoder codebase, this raises a
nondescript error.  The original issue is corrected, and the error message is made more
informative.  This is not the best place for this check, as it is done on a version of the
program source that users of the system are unlikely to have seen (the version that has been
rewritten with the new primitives, before they have been wrapped with `Invented ...`).
However, `minimal_inhabitant`, which produces these rewritings borrows most of its implementation
from functions which are not used purely to rewrite programs.  So a more thorough redesign
of the distribution of effort during beam searches and what follows would probably be
necessary to better distribute the efforts of this last bit of rewriting.