fix incorrect input to rewriting procedure
The rewriting procedure was being called such that it sought to replace instances
of `Invented refactoring` rather than `refactoring` in the inverse beta-reduced programs.
When this was not found (since the programs are not yet rewritten with `refactoring` as
an invented primitive), the inverse beta-reduced program resembled an ordinary program which
was out of beta-normal form.  This commit corrects the issue and improves on the nondescript
error message that was previously raised when this case is encountered.  This is not the
ideal place for this check, as it is done on a version of the program source that is likely
to be unfamiliar as it is rewritten with novel primitives not wrapped with `Invented`.
However, it would require a more thorough rewrite of all of the minimal inhabitant-related
functions to do it earlier.