refactor `Util`; move contents of `Transforms` to top-level `Antireduce` module
Because I don't think it is very clear what a "transform" is, I eliminated
that terminology from the codebase and move the transforms module to the
former `Commands` module, and made the combined module the top-level module
synonymous with the library.  In the process, I eliminated the unused `Util.range`
function, and refactored the `Util.singleton_list` function.  The assumptions
of this function were used only in relation to version spaces, and were justified
by a very specific quality of version tables.  So I created a function in `versions.ml`
that embodied that specific assumption, with specific error messages and a comment.
This covered five of the six uses of `Util.singleton_list`.  The remainder was made
clearer by replacement with an ordinary match statement.