redesign execution procedures for random search
The library is pivoting towards a use case more similar to the
dreaming functionality of dreamcoder: programs are enumerated
via random search ahead of time and only the shortest program
producing a given output is retained.  Programs are only generated
as a consequence of random search, and there is alwways only one
shortest program at a time.  During compression, different
primitives should be scored according to how they perform on a
frontier, as before, but all programs should be refactored with
the new primitive once it determined according to this criterion.

During the latter, I extracted a number of subroutines that, though
not necessarily used multiple times, were contributing to the bulkiness
of `compression_step` and `compress`.  To achieve the above aim of
refactoring all programs, but allowing the frontier to decide how
they are scored, the version table has all programs incorporated
into it, but the possible inventions are generated only from
examining the frontier.  As far as I can tell, this makes all programs
inside or outside of the frontier first-class citizens in terms of
permitting refactoring, while preserving the unique influence of
the frontier.