add missing if statement guarding removal of variables equivalent to goto
Because we are using CPS, if we do not make sure that each function acts on the
most recent state we might allow the equivalent of a goto (skipping the work done
to produce that state in favor of acting on an earlier one).  Depending on the
statefulness of the domain, this could also be hazardous--in addition to being
wasteful.

There was a missing if statement that prevented this removal of nonterminal unifying
environment variables if the requested type was not the state type of the domain.
This statement was added.