correct bug and potential bug in type checking/unification procedures
The current random enumeration is turning up programs that do not unify.  I
have carefully scrutinized those functions and the related functions in
`Dsl`, but it seems that they implement the same procedures as the DreamCoder
codebase, which I take to have functioned (at least at one point in time).
I have corrected to places where the antireduce procedure differed, one
small and one more likely to be the source of an issue as it used stateful
operations to produce the default value of a call to `Option.value_map`.