correct bug and potential bug in type checking/unification procedures
The current random enumeration is turning up programs that do not unify.  I
have carefully scrutinized those functions and the related functions in
`Dsl`, but they appear not to be doing anything differently from the
DreamCoder source.  I would like to believe there was not an error in
the DreamCoder source, so I corrected two small differences, one likely
less significant, but the other more significant (it put a stateful operation
as producing the default value of an `Option.value_map` call).