use `Obj.magic` coercions for domain-generic evaluation
We cannot prove that a `p : program` is isomorphic to a well-typed composition
of native OCaml functions without invoking the domain-specific implementations
of a given DSL.  Consequently, we cannot provide both domain-generic compression
and domain-generic execution of arbitrary programs.  The best compromise would
be to provide only domain-generic compression, allowing execution and enumeration
to be handled by downstream libraries in domain-specific ways.  This would allow
downstream libraries to use relatively straightforward techniques to build and
execute programs without them or antireduce needing implement any type inference
algorithms; OCaml's type inference could do this work.  However, this route would
require that both the inputs and oututs of antireduce's compression were type-annotated,
so that downstream libraries could easily incorporate invented primitives (and
detect bugs in antireduce manifesting through ill-typed inventions).  Tracking
types through the compression procedure so that they are preserved in the
resulting inventions without simply utilizing nonnative type inference to infer
them is not implemented, and would require significant work which cannot
be imitated from the DreamCoder system.  I cannot undertake this effort now, so
I am leaving domain-generic execution as an offering of antireduce and using
`Obj.magic` to dismiss the OCaml typechecker's well-founded by irrelevant issues
with this construction.

On a separate note, I have come to understand why the bug corrected and commented
on in commit `64ab488af8db6f6a4a09ed05f895b0ecee339396` was significant.  The
offending code was

```
    ...
    | FId ty_opt_ref ->
        Option.value_map !ty_opt_ref ~f:Fn.id
          ~default:(
            (let ty = Id !next_id_ref in
             incr next_id_ref ;
             ty_opt_ref := Some ty
             ty )
```

I changed this to

```
    ...
    FId ({contents= ty_opt} as ty_opt_ref) -> (
    match ty_opt with
    | Some ty ->
        ty
    | None ->
        let ty = Id !next_id_ref in
        incr next_id_ref ;
        ty_opt_ref := Some ty
        ty )
```

simply on the grounds that I recognized that the `default` branch of `Option.value_map`
may be executed even if `!ty_opt_ref` matched with `Some _`.  The specific reason this
caused an issue was that `ty_opt_ref` contained an inferred type of a type variable in
the signature of an invention.  By replacing it with a fresh type ID, I was undoing
the work of unification after it had succeeded, and therefore not appropriately propagating
constraints back to the enumeration functions.  This is what allowed a program to be
produced in which an integer was provided where a `graph_state` was required.  This
coincided with other bugs in the execution procedures due to previous doomed attempts to
provide domain-generic execution (which is not possible).

The above issues gave the appearance that fast types and fast unification were not working
properly, when in fact they mostly were, the constraints resulting from their proper
execution were simply being erased.