remove domain-general program execution
The use of `Obj.magic` seemed to be having puzzling side effects
and was very difficult to debug, so I removed it in favor of
offering a more limited form of program synthesis which delegates
to downstream libraries for native OCaml execution.  These libraries
must define a correspondence between GADTs defining types and terms
and the non-GADTs defining manually typechecked types and terms,
however they do not have to replicate the work of type inference,
nor do we have to worry about preserving type information through
inverse beta reduction (we maintain the ability to infer it manually
for all new primitives).