fix bugs in `decompress_generic`
In order to fully type a lambda calculus program and all its subexpressions,
you need to first instantiate type variables for all unknown types such that
substructure is properly maintained (utilize a variable context of types used
at the appropriate places and unify the types of function parameters and their
arguments); you must then unify these types recursively with the top-level
requested type.  Once you have done this (and returned the type context in which
these unifications were performed) you may apply the appropriate substitutions
to translate every type ID on every subexpression to its inferred (monomorphic)
type according to that type context.  The instantiation and unification could
perhaps be performed in a single step, but they are easier to reason about
separately.  However, the instantiation/unification and application of the
substitutions resulting from type inference cannot be collapsed into a single
pass, as we must process both functions and their arguments to infer the type
of all subexpressions but we must logically process one before the other.  If we
attempted to combine type inference and depositing of type annotations into one
pass, the type annotations we produce may not be monomorphic.