ignore types of invented primitives for program comparison and equality
To create a parser that does not depend on the specific invented primitives
in a DSL in order to parse programs that may contain invented primitives,
the types of invented primitives are inferred during parsing.  This
can result in invented primitives in parsed programs not having strictly
identical types to those of the invented primitives in a DSL.  This was
messing with program equality, as `equal_dc_type` does not perform any
form of unification.  However, to prevent programs from registering as
equal even if they come from different domains, we would prefer to have
program equality depend somewhat on the types of the program elements,
rather than on just their names (which could fairly easily clash across
different domains).  To allow the parser to work generically across
DSLs developed within a certain domain, and to preserve the efficiency
of the program equality operation, I ignore the type of invented primitives
for the purpose of program equality/comparison only.  Thus, types of
primitives are still compared, and this offers some extra assurance
that programs from different domains will not register as equal, but
types inferred during parsing are not directly compared.  This sacrifices
nothing in terms of the integrity of the program equality operation,
as two invented primitives consisting of the same configuration of
base primitives, all having the same type, necessarily have unifiable types.