select new primitives according to size with tie-breaking done by score reduction
Previously, new primitives were selected according to which most compressed the frontier.
However, I realized that while this makes sense when we have preferences about the semantics
of programs, it doesn't make sense otherwise.  In DreamCoder, we would know that if
task solutions exhibit optimal substructure w.r.t. each other, program enumeration will be
more likely to find correct solutions to new tasks if it uses inventions that optimally
compress existing solutions.  In the PSN setup, program enumeration is performed by training
a neural network which is most likely to favor a program's output as its internal representation
(and therefore include it in the frontier) if it bears a useful resemblance to the outputs
of other programs.  This is most likely if the invented primitives represent minimal departures
departures from how existing programs are written.  So while we still want the most relevant
primitives, it is more important that invented primitives are small.  This helps to ensure
that outputs of programs synthesized using this primitive are more likely to influence the PSN
and to be included in the frontier, which then encourages hierarchically-structured primitives
to be produced.

tl;dr: when we choose the most compressive primitive for a frontier that lacks diversity of
structure, and which is composed by a mechanism that likely favors gradual complexification
of programs, we contribute to ungainly primitives, which are not effective in producing
a large variety of programs and program outputs.