correct bug causing programs not to be selected by size
By discarding all but the `top_i` inventions before sorting by
size, I was preventing the smallest invented primitives from
actually being selected.  In the process of examining this,
I noticed that the two-stage scoring implemented in `compression_step`
seemed pointless.  Upon closer examination, I saw that it had been
rendered pointless by haphazard removal of the reestimation of DSL
parameters, which is an expensive enough step to justify performing it
only for select approximately optimal inventions.  In the absence of
this step, the two stage scoring was basically redundant, just more
exact and incorporating other penalties and criteria.  For now,
I decided simply to utilize the cost as calculated with `beams_and_costs`
to break ties amongst primitives of the smallest `n_invention_sizes`
sizes, and to only calculate the exact score using the rewritten
frontier of the single selected primitive.  I also removed the
invented primitive/DSL size penalties from the score.

Additionally, measurements of invented primitive sizes (and the
frontiers/DSLs constituted from them) are important for determining
whether invented primitives are effective in producing a useful
diversity of program outputs, and whether this diversity of outputs
is being utilized by a PSN.  To facilitate these measurements, I
defined the `absolute_size_of_program` to capture the size of
programs and DSLs in terms of base primitives.  I modified the
`dsl` type, the exploration procedures, and the post-compression
refactoring procedures to make this measurement (which I refer to
as the "mass" of primitives and DSLs/programs composed with them)
available in the artifacts produced by the library.