revise program search strategy
In the previous commit, I removed the backtracking that was causing
search to be depth-first (though reinitialized after each program
was found) because it was causing very long search times.
However, this purely random enumeration was not performing very
well in terms of program diversity.  I think that partly, this was
due to the difficulty of doing a depth-first search in a very tightly
depth-limited search space.  I also believe that the relatively
shallow search space was severely limiting the size of graphs that
were being discovered through exploration.  To address these problems,
I instead implemented a procedure very similar to that used in
DreamCoder.

In DreamCoder, best-first search is conducted until a certain number
of candidates are discovered, then depth-first search is conducted
from these starting points.  Unlike in DreamCoder, however, we cannot
rank programs in terms of their likelihood to accomplish a task, since
we have no fixed set of tasks.  This makes it difficult to choose exactly
how to trade off breadth-first for depth-first search.  Instead of choosing
this cutoff arbitrarily, we iteratively deepen the breadth-first search,
and perform depth first search on each candidate until the first solution
is found (though iterative deepening can also be disabled).  Programs are
evaluated and stored as soon as they are found, so the `exploration_time`
bound should be adhered to quite closely.

An absolute upper bound on program size is still provided which even
depth-first search will not exceed.  Additionally, one can specify
how many backtracking points are included in depth-first search, with
these points being roughly evenly-spaced.  This allows one to perform
ordinary DFS as well as abridged forms which simply fail upon reaching
the DFS depth limit.