restore original DreamCoder enumeration and compression functionality
I realized that PSNs may operate in conjunction with any program
synthesis system that is both parametrized by a set of I/O examples
and produces I/O examples when operated.  This describes DreamCoder,
which is of course parametrized by a body of I/O examples in a
domain, and which produces I/O examples upon executing synthesized
programs.  The key difference is that in the original formulation
of DreamCoder, the set of I/O examples that parametrizes program
synthesis and the set of I/O examples we want to produce from
operating the system are the same.  This is what obligates the
use recognition models to bias search towards programs that produce
the desired outputs, even before there is evidence to suggest
which program structures one should be biased towards to accomplish
this.  Because PSNs offer an open-ended way to evaluate any set of
I/O examples for their usefuleness in accomplishing a downstream
task, there is no requirement that the set of I/O examples that
parametrize program synthesis at point A be the same as that which
it generates--by whatever definition--thereafter; nor that the program
synthesis system not be _re_parametrized by a new set of I/O examples
later.  This allows us to parametrize our distributional search over
programs by using exact data on which programs _produced_ a set of
I/O examples, eliminating the need for recognition models. It also
loosens the requirements on how the programs (and the I/O examples
they generate) should complexify as the program synthesis system
operates: a system can accomplish this in any way that brings empirical
success to the downstream PSN.

The consequences of this are that any programming by example (PbE)
system that searches over increasingly large search spaces can be
paired with a PSN.  DreamCoder is not unique in fitting this bill;
in fact, all of its constituent functionalities appear to have been
improved upon in more recent work (particularly: distributional
search and compression).  The only lasting contribution of DreamCoder
is the general tactic of combining distributional search with
compression of the search space.  Consequently, this commit recreates
the distributional search and compression behavior of the DreamCoder
system, as doing otherwise would probably involve no longer using this
codebase in favor of using newer implementations created by more recent
research work.

tl;dr: originally, distributional search functionality was omitted because
I conceived of the PSN as operating on representations of programs,
not I/O examples themselves.  The PSN was to fulfill the roles of I/O
example curator, recognition model, and enumerator at once.  I saw then
that the PSN was not succeeding at learning any relationship between
program syntax and program semantics, so I removed the burden of
enumerating programs from the PSN and had it work only with precomputed
outputs.  I still assumed, though, that the feedback the PSN offered on
which I/O examples were useful for the downstream task was relevant only
to guiding compression, as I had focused so little on distributional
search as a mechanism.  I realize now that this feedback may be used in
any manner by any PbE system, and may inform any of its constituent
parts.