relax `strip_n_abstractions` to reducing n or fewer abstractions to acommodate lack of DreamCoder normal form
DreamCoder's normal form expects that programs always use lambdas to pass arguments
even when the underlying primitive immediately unifies with the requested type.  This
translated to a requirement that `strip_n_abstractions` throw an error if it could not
strip away exactly as many abstractions as there were arguments to a type.  This caused
a problem when trying to evaluate the likelihood of a program not in this form.