make factorization agnostic to presence of DreamCoder normal form
The `factorize` function relied on the assumption that types should be
deconstructed until their return type is reached, at which point
unifiable candidates could be generated to match the return type; along
the way, the DreamCoder normal form would have it that there was one
abstraction to peel off for each argument of a type that was deconstructed.
The previous commit silenced the error that was raised when there was
no such abstraction to strip away, but left the remaining behavior intact.
This commit makes the `factorize` function fully agnostic to the presence
of this normal form by deconstructing the current requested type only
insofar as there are abstractions corresponding to the arguments indicated
by that type.  If there are no such abstractions, `factorize` now expects
to be able to unify the entire type request with some suffix of the function
of the application that matches it.