fix bug in mechanism for aligning types of primitives with requested types
Currently, primitives' types are unified with requested types by
trying to unify the smallest return types of the respective signatures.
This imposes the assumption that a requested type `constructor_1 -> t0`
corresponds to programs whose type contains one right-associative arrow.
That is, it unifies with `constructor_1 -> constructor_2`, and not with
any other types such that `t0` itself contains an arrow.  To get programs
with such types, the requested type must dictate the appropriate arity
with:

    `constructor_1 -> t0 -> t0 -> t1`

This could possibly produce different results than the request

    `constructor_1 -> t0`

even though in OCaml, say, these are equivalent.  This is also how
the DreamCoder codebase handles this circumstance.

I suspect, but don't know how to prove, that this practice makes the
way that  invented primitives may be used a bit more limited, as when
a type variable is inferred to represent an index, there may be multiple
arities of function that could be substituted for it.  However, under
this definition of unification, only one such arity is considered.