revise factorization procedure and pcfg-of-dsl transformation to acommodate stitch invented primitives
As discussed in previous commits, when investigating an issue with factorization
that was arising due to the introduction of stitch-invented primitives, I found
that the unification procedure used for fast types seemed less than general, in
that arrows were interpreted as specifying the exact arity of the functions they
identified (contrary to their generally understood meaning).  This seemed worth
correcting to me, if only so that the treatement of invented primitives in the
`pcfg_of_dsl` could be made more precise.  Indeed, this fix uncovered existing
problems with the `pcfg_of_dsl` transformation, mainly stemming from the fact
that the previous implementation of this transformation was not fully unifying
the types of the programs it was uncovering before turning them into productions
of the grammar.  The core idea of enumerating unifying combinations of expressions
without recursing on the previously encountered types was correct, however when
doing this it was important to make sure that each such partial derivation was
paired with its complete type context.  This allowed partial derivations to be
fully unified so that all productions gleaned from them were monomorphic to the
extent required by the request.  I realized that the same issue was at play in
the `factorize` function.  We want to calculate the likelihood of choosing a
particular primitive from the list of possible primitives; we don't want to
count  the multiple ways that a primitive could potentially be unified with
the requested type (if such multiple ways exist).  This prompted me to make
likelihood summaries immutable (Map-based) to facilitate factorizing according
to all possible unifications, but I realized that it was simpler to reuse the
ability to create a fully-typed program.  By treating invented primitives as
primitives within a fully-typed program and then converting them back to invented
primitives when performing the factorization, we could eliminate the ambiguity in
how any subexpression of the program should be typed beyond that inherent in the
requested type; we could then simply traverse the fully typed tree, caring only
about the programs that unify rather than their context or or parameters.``

Along the way, I started being irked by names such as `empty_type_context`
and `string_of_program` so I refactored everything to make greater use of modules.
There are now many more files in the codebase, but pretty much the same amount of
code in smaller chunks.  Overall, I think this makes the code more readable as
types are captured in more transparent ways than they were when only suggested
by naming conventions.  It also led to a solution for the record name clashing
problem, which is to use `ppx_fields_conv` to create field extracting functions.
These functions are convenient, of course, but also narrow the type of a record
that might otherwise not be distinguishable based only on how its fields are
accessed.  Using a particular module's accessor solves this problem, and seems
more robust and refactorable than adding type annotations in arbitrary locations.