make hashing of programs sensitive to invention names; update "stitch_program" JSON field
For safety in preventing name clashes, it is better to produce hashes of programs
that are fully sensitive to the name-implementation correspondence established for
invented primitives; this can be done by hashing based on the ``Combined` program
serialization format.  Also, the "stitch_program" field of the representations
JSON was not being updated.  This was causing a number of issues downstream with
stitch compression via Python as Python code reads in this representation of the
frontier when calling the stitch compression tool.