support "multikey" exploration in addition to previous ("unikey") exploration
Previously, it was assumed that programs/outputs were uniquely identified with
a single key (they could be hashed).  To support domains where this assumption
may not hold, we allow for both a primary key (a hash) and a secondary key
supporting equality which can be used to do linear search among those programs/
outputs not differentiated by a particular primary key.  In the extreme case
of lacking any primary key (supporting only linear search), a downstream domain
may have a single primary key for all possible programs/outputs.  This change
is meant to address circumstances in which total ordering of a class of program
outputs is not available.