restore original `type_size_limit`; correct looping recursion bug in `Partial_derivation.find`
If a primitive has two type variables, then a situation can arise in which
unifying with its signature creates a request which is unifiable with a
previously encountered type during the search, but is not exactly equal.
This opens the door for looping recursion.  Such types, when they arise,
should have no constraints (according to the current type context) that
prevents them with unifying with one of the previously-seen nonterminals.
This should prevent the looping recursion without preventing recursion on
types for which there are constraints making them genuinely novel.