add module offering structural type equality; integrate with `Partial_derivation.find`
I believe that using lists rather than sets to represent `seen_nts` and
`completed_nts` could have significant performance consequences, or at
least it does in the pathological case of looping recursion I am debugging
at present.  To address this, I created a module, `Structural_type`, that
assumes the presence of an optional context used for unification-based
equality and defaults to `Type.compare` for the other cases of comparison
if this is not available or if unified types are unequal.  This allows a
`Structural_type.t` to be stored in an ordered data structure, provided
that they are tested for membership and unified prior to insertion using
the most up-to-date type context.