track primitive to primitive transitions during `Pcfg.of_dsl` transformation
I have yet to find a way to prevent costly unbounded recursion on increasingly
large but uninstantiable types through a criterion that examines the requested
type alone.  However, I remembered that in the DeepSynth implementation of heap
search, nonterminals also include information about which argument of which
primitive is being supplied when generating a derivation.  Since there is never
anything new to be discovered, in terms of the completeness of the PCFG, by
exploring the same transition (primitive _i_, arg j -> primitive k) more than
once, it seems that we should be able to prevent types from growing infinitely
larger simply by tracking and ruling out redundant recursive calls under this
criterion.